# server/openapi.yaml
...
openapi: 3.1.0
info:
  title: Fake Stack Overflow API
  version: 1.0.0
  description: OpenAPI 3.1 specification for the Fake Stack Overflow backend server.
servers:
  - url: /api
    description: The API server endpoint
paths:
...
  /question/getQuestion:
    get:
      tags:
        - Questions
      summary: Retrieve questions filtered by order, search, and asker.
      parameters:
        - in: query
          name: order
          schema:
            $ref: '#/components/schemas/OrderType'
          required: true
          description: The order in which to sort the questions.
        - in: query
          name: search
          schema:
            type: string
            default: ''
          required: false
          allowEmptyValue: true
          allowReserved: true # removes the requirement that search parameter should be URL encoded.
          description: Search string for questions, can include tags or keywords.
        - in: query
          name: askedBy
          schema:
            type: string
          required: false
          description: Username of the user who asked the question.
      responses:
        '200':
          description: Successfully retrieved filtered list of questions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PopulatedQuestion'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /question/getQuestionById/{qid}:
    get:
      tags:
        - Questions
      summary: Retrieve a single question by its ID, incrementing its view count.
      parameters:
        - in: path
          name: qid
          schema:
            type: string
            format: object-id
          required: true
          description: The unique ID of the question.
        - in: query
          name: username
          schema:
            type: string
          required: true
          description: The username of the viewer to track views.
      responses:
        '200':
          description: Successfully retrieved question details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PopulatedQuestion'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /question/addQuestion:
    post:
      tags:
        - Questions
      summary: Add a new question.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewQuestionBody'
      responses:
        '200':
          description: Question successfully added and populated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PopulatedQuestion'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /question/upvoteQuestion:
    post:
      tags:
        - Questions
      summary: Upvote a question.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoteBody'
      responses:
        '200':
          description: Vote successfully processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoteResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /question/downvoteQuestion:
    post:
      tags:
        - Questions
      summary: Downvote a question.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoteBody'
      responses:
        '200':
          description: Vote successfully processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoteResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /question/getCommunityQuestions/{communityId}:
    get:
      tags:
        - Questions
      summary: Retrieve questions for a specific community.
      parameters:
        - in: path
          name: communityId
          schema:
            type: string
            format: object-id
          required: true
          description: The unique ID of the community.
      responses:
        '200':
          description: Successfully retrieved list of community questions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PopulatedQuestion'
        '500':
          $ref: '#/components/responses/ErrorResponse'

  # --------------------
  # ANSWER CONTROLLER
  # --------------------
...
  /user/signup:
    post:
      tags:
        - Users
      summary: Create a new user account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSignupBody'
      responses:
        '200':
          description: User successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /user/login:
    post:
      tags:
        - Users
      summary: Log a user in.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginBody'
      responses:
        '200':
          description: User successfully logged in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /user/resetPassword:
    patch:
      tags:
        - Users
      summary: Reset a user's password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginBody'
      responses:
        '200':
          description: User password successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /user/getUser/{username}:
    get:
      tags:
        - Users
      summary: Retrieve a user by username.
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: The username of the user to retrieve.
      responses:
        '200':
          description: Successfully retrieved user details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /user/getUsers:
    get:
      tags:
        - Users
      summary: Retrieve a list of all users.
      responses:
        '200':
          description: Successfully retrieved list of all users.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /user/deleteUser/{username}:
    delete:
      tags:
        - Users
      summary: Delete a user by username.
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: The username of the user to delete.
      responses:
        '200':
          description: User successfully deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /user/updateBiography:
    patch:
      tags:
        - Users
      summary: Update a user's biography.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBiographyBody'
      responses:
        '200':
          description: User biography successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ErrorResponse'
components:
  schemas:
...
    OrderType:
      type: string
      enum:
        - newest
        - unanswered
        - active
        - mostViewed
      default: 'newest'
    VoteResponse:
      type: object
      properties:
        msg:
          type: string
        upVotes:
          type: array
          items:
            type: string
        downVotes:
          type: array
          items:
            type: string

    # --------------------
    # QUESTION SCHEMAS
    # --------------------
    VoteBody:
      type: object
      required:
        - qid
        - username
      properties:
        qid:
          type: string
          description: Unique identifier of the question.
        username:
          type: string
          description: Username of the user casting the vote.
    TagBody:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          minLength: 1
          description: The name of the tag.
        description:
          type: string
          minLength: 1
          description: A brief description of the tag's purpose or usage.
    NewQuestionBody:
      type: object
      required:
        - title
        - text
        - tags
        - askedBy
        - askDateTime
      properties:
        title:
          type: string
          minLength: 1
        text:
          type: string
          minLength: 1
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagBody'
          minItems: 1
          description: Array of Tag objects, each requiring a name and description.
        askedBy:
          type: string
          minLength: 1
        askDateTime:
          type: string
          format: date-time
        community:
          type: string
          format: object-id
          nullable: true
    PopulatedQuestion:
      type: object
      description: A fully populated question object with embedded tags, answers, and comments.

    # --------------------
    # ANSWER SCHEMAS
    # --------------------
...
    UserSignupBody:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          nullable: true
          description: Optional user email.
        biography:
          type: string
          default: ''
          description: Optional user biography.
    UserLoginBody:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
    UpdateBiographyBody:
      type: object
      required:
        - username
        - biography
      properties:
        username:
          type: string
          minLength: 1
        biography:
          type: string
    User:
      type: object
      description: A user profile object.
...
  responses:
...
    ErrorResponse:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: The error message returned by the server.
    ValidationError:
      description: Bad Request / Validation Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: The error message from express-openapi-validator.
              errors:
                type: array
                description: Detailed list of validation errors.